test-embedding/api-vector/requirements.txt
fastapi 
uvicorn 
python-dotenv 
psycopg2 
requests

psycopg2-binary 
tiktoken
transformers 
torch